+++ Tests/LegacyUnit/BackEnd/TcaTest.php
@@ -51,9 +51,7 @@ class tx_realty_BackEnd_TcaTest extends \Tx_Phpunit_TestCase
      */
     public function getDistrictsForCityContainsEmptyOption()
     {
-        $result = $this->subject->getDistrictsForCity(
-            ['row' => ['city' => 0]]
-        );
+        $result = $this->subject->getDistrictsForCity(['row' => ['city' => 0]]);
 
         self::assertContains(['', 0], $result['items']);
     }
@@ -61,26 +59,19 @@ class tx_realty_BackEnd_TcaTest extends \Tx_Phpunit_TestCase
     /**
      * @test
      */
-    public function getDistrictsForCityReturnsDistrictsForCityOrUnassigned()
+    public function getDistrictsForCityReturnsDistrictsForCity()
     {
-        $city = new tx_realty_Model_District();
+        $city = new \tx_realty_Model_District();
         $city->setData(['uid' => 2, 'title' => 'Kreuzberg']);
-        $cities = new Tx_Oelib_List();
+        $cities = new \Tx_Oelib_List();
         $cities->add($city);
 
         /** @var tx_realty_Mapper_District|PHPUnit_Framework_MockObject_MockObject $mapper */
-        $mapper = $this->getMock(
-            \tx_realty_Mapper_District::class,
-            ['findAllByCityUidOrUnassigned']
-        );
-        $mapper->expects(self::once())
-            ->method('findAllByCityUidOrUnassigned')->with(42)
-            ->will(self::returnValue($cities));
-        Tx_Oelib_MapperRegistry::set('tx_realty_Mapper_District', $mapper);
+        $mapper = $this->getMock(\tx_realty_Mapper_District::class, ['findAllByCityUid']);
+        $mapper->expects(self::once())->method('findAllByCityUid')->with(42)->willReturn($cities);
+        \Tx_Oelib_MapperRegistry::set(\tx_realty_Mapper_District::class, $mapper);
 
-        $result = $this->subject->getDistrictsForCity(
-            ['row' => ['city' => 42]]
-        );
+        $result = $this->subject->getDistrictsForCity(['row' => ['city' => 42]]);
 
         self::assertContains(['Kreuzberg', 2], $result['items']);
     }